rename: 프로필 변경 api 네이밍 변경
